backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1

# Data manipulation
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
tqdm==4.66.1

# Export functionality
reportlab==4.0.7
PyPDF2==3.0.1

# Mock ML (lightweight placeholders)
scikit-learn==1.3.2

# API utilities
marshmallow==3.20.1
flask-marshmallow==0.15.0


